chore(polyfills-loader): update snapshots
