+++ packages/create/test/snapshots/fully-loaded-app/.storybook/main.js
@@ -1,5 +1,5 @@
 module.exports = {
-  stories: ['../components/*/stories/*.stories.{js,mdx}'],
+  stories: ['../**/stories/*.stories.{js,mdx}'],
   esDevServer: {
     // custom es-dev-server options
   },
