--- packages/create/test/snapshots/fully-loaded-app/components/page-main/page-main.js
@@ -1,3 +1,0 @@
-import { PageMain } from './PageMain.js';
-
-window.customElements.define('page-main', PageMain);
