--- packages/create/test/snapshots/fully-loaded-app/components/page-one/page-one.js
@@ -1,3 +1,0 @@
-import { PageOne } from './PageOne.js';
-
-window.customElements.define('page-one', PageOne);
