+++ packages/create/test/snapshots/fully-loaded-app/index.html
@@ -1,5 +1,6 @@
 <!doctype html>
 <html lang="en">
+
 <head>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0, viewport-fit=cover" />
@@ -7,7 +8,8 @@
   <base href="/">
 
   <style>
-    html, body {
+    html,
+    body {
       margin: 0;
       padding: 0;
       font-family: sans-serif;
@@ -16,10 +18,11 @@
   </style>
   <title>scaffold-app</title>
 </head>
-<body>
-  <scaffold-app title="Hello owc world!"></scaffold-app>
 
-  <script type="module" src="./components/scaffold-app/scaffold-app.js"></script>
+<body>
+  <scaffold-app></scaffold-app>
+
+  <script type="module" src="./src/scaffold-app.js"></script>
 </body>
 
 </html>
