+++ packages/create/test/snapshots/fully-loaded-app/src/open-wc-logo.js
