+++ packages/create/test/snapshots/fully-loaded-app/src/scaffold-app.js
