+++ packages/polyfills-loader/test/snapshots/inject-polyfills-loader/module-and-legacy.html
@@ -1,4 +1,4 @@
-<html><head><link rel="preload" href="./app.js" as="script" crossorigin="anonymous"></head><body><div>before</div>
+<html><head></head><body><div>before</div>
       <script type="module" src="./app.js"></script>
       <div>after</div>
     <script>(function () {
