+++ packages/polyfills-loader/test/snapshots/inject-polyfills-loader/no-polyfills-no-legacy.html
@@ -1,4 +1,4 @@
-<html><head><link rel="preload" href="./app.js" as="script" crossorigin="anonymous"></head>
+<html><head></head>
 
       <body>
         <div>Hello world</div>
