+++ packages/polyfills-loader/test/snapshots/inject-polyfills-loader/systemjs-polyfill-importmap-modern.html
@@ -1,6 +1,6 @@
 <html><head>
         <script type="importmap">{ "imports": { "foo": "bar" } }</script><script type="systemjs-importmap">{ "imports": { "foo": "bar" } }</script>
-      <link rel="preload" href="./app.js" as="script" crossorigin="anonymous"></head>
+      </head>
       <body><div>before</div>
       <script type="module" src="./module-a.js"></script>
       <div>after</div>
